Many tweaks to reduce bitching by clang static analysis.
ESL_ALLOC() and _REALLOC() of 0 size now return eslEMEM error, not
NULL. Having them return NULL confuses clang.

ESL_ANALYZER_NORETURN macro added: to tell clang that esl_fatal() is a
fatal error handler that never returns.




git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@953 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ esl_randomseq.c
@@ -1679,7 +1679,7 @@ utest_XShufflers(ESL_RANDOMNESS *r, int L, int K)
   /* allocations */
   ESL_ALLOC(dsq, sizeof(ESL_DSQ) * (L+2));
   ESL_ALLOC(ds2, sizeof(ESL_DSQ) * (L+2));
-  ESL_ALLOC(p,   sizeof(double)  * K);
+  ESL_ALLOC(p,   sizeof(float)   * K);
   if (composition_allocate(K, &m1, &di1) != eslOK) esl_fatal(logmsg);
   if (composition_allocate(K, &m2, &di2) != eslOK) esl_fatal(logmsg);
 
@@ -1789,7 +1789,7 @@ utest_XMarkovs(ESL_RANDOMNESS *r, int L, int K)
   /* allocations */
   ESL_ALLOC(dsq, sizeof(ESL_DSQ) * (L+2));
   ESL_ALLOC(ds2, sizeof(ESL_DSQ) * (L+2));
-  ESL_ALLOC(p,   sizeof(double)  * K);
+  ESL_ALLOC(p,   sizeof(float)   * K);
   if (composition_allocate(K, &m1, &di1) != eslOK) esl_fatal(logmsg);
   if (composition_allocate(K, &m2, &di2) != eslOK) esl_fatal(logmsg);
